DIGITAL CALCULATOR
This project is particularly GUI-driven, making it easy to execute. This project will be a menu-driven program in which the outputs will be given by the user. It assists to strengthen some of the fundamental concepts of C++, particularly relating data types and decision-making statements.

The project requires validation for the inputs so that the calculation only takes place if the inserted data is numerical by nature. Furthermore, there’s some concept of data handling of a particular size. The larger the number, the larger the data type required to handle it. Tackling it properly will also need some skills that will go a long way in enabling one to secure an internship/job in the future.